+++ src/lib/components/modes/JSONEditorRoot.svelte
@@ -40,6 +40,7 @@
   export let mainMenuBar: boolean
   export let navigationBar: boolean
   export let statusBar: boolean
+  export let askToFormat: boolean
   export let escapeControlCharacters: boolean
   export let escapeUnicodeCharacters: boolean
   export let flattenColumns: boolean
@@ -183,12 +184,13 @@
     }
   }
 
-  export function scrollTo(path: JSONPath): void {
+  export function scrollTo(path: JSONPath): Promise<void> {
     if (refTreeMode) {
       return refTreeMode.scrollTo(path)
+    } else if (refTableMode) {
+      return refTableMode.scrollTo(path)
     } else {
       // TODO: implement scrollTo for text mode
-
       throw new Error(`Method scrollTo is not available in mode "${mode}"`)
     }
   }
@@ -227,9 +229,9 @@
     }
   }
 
-  export function refresh() {
+  export async function refresh(): Promise<void> {
     if (refTextMode) {
-      refTextMode.refresh()
+      await refTextMode.refresh()
     } else {
       // nothing to do in tree or table mode (also: don't throw an exception or so,
       // that annoying having to reckon with that when using .refresh()).
@@ -246,6 +248,7 @@
     {tabSize}
     {mainMenuBar}
     {statusBar}
+    {askToFormat}
     {escapeUnicodeCharacters}
     {parser}
     {validator}