+++ src/lib/components/modes/tablemode/ColumnHeader.svelte
@@ -4,13 +4,12 @@
   import type { JSONPath } from 'immutable-json-patch'
   import type { SortedColumn } from '$lib/types.js'
   import { SortDirection } from '$lib/types.js'
-  import { stringifyJSONPath, stripRootObject } from '../../../utils/pathUtils.js'
+  import { stringifyJSONPath } from '../../../utils/pathUtils.js'
   import Icon from 'svelte-awesome'
   import { faCaretDown, faCaretUp } from '@fortawesome/free-solid-svg-icons'
   import { isEmpty, isEqual } from 'lodash-es'
-  import { SORT_DIRECTION_NAMES } from '$lib/constants.js'
+  import { MAX_HEADER_NAME_CHARACTERS, SORT_DIRECTION_NAMES } from '$lib/constants.js'
   import { truncate } from '$lib/utils/stringUtils.js'
-  import { MAX_HEADER_NAME_CHARACTERS } from '$lib/constants.js'
 
   export let path: JSONPath
   export let sortedColumn: SortedColumn | undefined
@@ -18,7 +17,7 @@
   export let onSort: (sortedColumn: SortedColumn) => void
 
   // TODO: improve truncating of long column names when they are a deeply nested path: the last item from the path should be visible, and halfway the path is least interesting
-  $: columnName = !isEmpty(path) ? stripRootObject(stringifyJSONPath(path)) : 'values'
+  $: columnName = !isEmpty(path) ? stringifyJSONPath(path) : 'values'
 
   $: sortDirection = isEqual(path, sortedColumn?.path) ? sortedColumn.sortDirection : undefined
   $: sortDirectionName = SORT_DIRECTION_NAMES[sortDirection]