+++ src/lib/plugins/validator/createAjvValidator.ts
@@ -45,9 +45,18 @@ export function createAjvValidator(options: AjvValidatorOptions): Validator {
   let ajv = createAjvInstance(options)
   if (options.onCreateAjv !== undefined) {
     ajv = options.onCreateAjv(ajv) || ajv
+
+    // validate whether ajv is configured correctly (this is needed to enhance error messages)
+    if (ajv.opts.verbose === false) {
+      throw new Error('Ajv must be configured with the option verbose=true')
+    }
   }
 
   const validateAjv = ajv.compile(options.schema as Schema)
+
+  if (validateAjv.errors) {
+    throw validateAjv.errors[0]
+  }
 
   return function validate(json: JSONValue): ValidationError[] {
     validateAjv(json)