+++ svelte.config.js
@@ -9,18 +9,6 @@ const config = {
 
   kit: {
     adapter: adapter()
-  },
-
-  package: {
-    dir: 'package',
-    files: (filepath) => {
-      // ignore test and snapshot files
-      if (filepath.endsWith('.test.ts') || filepath.includes('__snapshots__')) {
-        return false
-      }
-
-      return true
-    }
   }
 }